% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_braket.R
\name{geom_braket}
\alias{geom_braket}
\title{geom_braket: draw a bracket between two x positions with a label}
\usage{
geom_braket(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  tip_length = unit(0.02, "npc"),
  inherit.aes = TRUE,
  na.rm = FALSE,
  show.legend = NA
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}}

\item{data}{The data to be displayed in this layer}

\item{stat}{The statistical transformation to use on the data}

\item{position}{Position adjustment}

\item{...}{Other arguments passed on to methods}

\item{tip_length}{Vertical extent of bracket tips (default: \code{unit(0.02, "npc")})}

\item{inherit.aes}{Whether to inherit aesthetics from ggplot default}

\item{na.rm}{Remove missing values?}

\item{show.legend}{Should this layer be included in legends?}
}
\value{
A ggplot2 layer
}
\description{
Draws a horizontal bracket between \code{x} and \code{xend} at height \code{y}, with a label centered above it.
}
\examples{
library(ggplot2)
df <- data.frame(x = 1, xend = 3, y = 5, label = '*p* < 0.05')
ggplot() +
  geom_point(aes(1:5, 1:5)) +
  geom_braket(data = df, aes(x = x, xend = xend, y = y, label = label)) +
  ylim(0, 6)
}
