% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jgg-package.R
\docType{package}
\name{jgg-package}
\alias{jgg}
\alias{jgg-package}
\title{jgg: Small ggplot2 Extensions}
\description{
Extensions to ggplot2 including brackets and annotations.
}
\author{
\strong{Maintainer}: Jan Gleixner \email{jan@example.com}

}
